+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Changed
+
+- Clicking a map panel button for the panel already on screen now closes it, instead of only the header's X doing that. (#3317)
+
 ## [1.11.0] - 2026-08-02, Berlin
 
 ### Added